+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.24.0 (2026-mm-dd)
 
 - **ENH: Adding the support of `Satellogic` (`Aleph-1`) constellation** [#229](https://github.com/sertit/eoreader/issues/229)
+- **ENH: Add fallback for TSX/TDX extent when `SUPPORT/GEARTH_POLY.kml` is missing** [#290](https://github.com/sertit/eoreader/issues/290)
 - FIX: Fix border effect (extent replacing footprint) when computing the extent before the footprint for `Umbra` products
 - FIX: Ensure outputing quicklook paths as strings
 - FIX: Get correct buffer for windows in geographic CRS when subsetting SAR data